Initial Split Operation
