+++ .github/workflows/python-package.yml
@@ -70,4 +70,4 @@ jobs:
           SA_CLIENT_ID: ${{ secrets.SA_CLIENT_ID }}
           SA_CLIENT_X509_CERT_URL: ${{ secrets.SA_CLIENT_X509_CERT_URL }}
       run: |
-        python baseball-japan-npb-crawler.py upload:$UPLOAD
+        python baseball-japan/main.py upload:$UPLOAD
